tests/approved/audio/audio_loop_reflect.html

<!DOCTYPE html>
<html>
  <head>
    <title>Audio Test: audio_loop_true_loop_present</title>
    <link rel="author" title="Intel" href="http://www.intel.com" />
    <link rel="help" href="http://dev.w3.org/html5/spec/single-page.html#dom-media-loop" />
    <meta name="flags" content="" />
    <meta name="assert" content="Check that the loop reflects the loop content attribute in the audio element" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <audio id="m" loop>The user agent doesn't support media element.</audio>
    <script type="text/javascript">
        var media = document.getElementById("m");
        
        test(function() {
            assert_true(media.loop, "media.loop of value");
        }, "Check if audio.loop returns true when the loop attribute is present in the audio element");
        
        test(function() {
            media.loop = false;
            assert_false(media.loop, "media.loop of value");
        }, "Check if audio.loop returns false after the loop attribute is set to false");
        
        test(function() {
            media.loop = true;
            assert_true(media.loop, "media.loop of value");
        }, "Check if audio.loop returns true after the loop attribute is set to true");
    </script>
  </body>
</html>
